+++ src/components/drag-and-drop/draggable-context/draggable-context.js
@@ -1,6 +1,8 @@
 import React from 'react';
 import { DragDropContext } from 'react-dnd';
+import PropTypes from 'prop-types';
 import HTML5Backend from 'react-dnd-html5-backend';
+import ItemTargetHelper from './../../../utils/helpers/dnd/item-target';
 
 /**
  * A draggable context component
@@ -9,24 +11,26 @@ import HTML5Backend from 'react-dnd-html5-backend';
  *
  * In your file
  *
- *   import DraggableContext from 'carbon/lib/components/drag-and-drop'
+ *   import { DraggableContext, WithDrop, WithDrag } from 'carbon/lib/components/drag-and-drop'
  *
- * A draggable context is used to define an area in the page
- * where drag and drop can be used on one or more elements:
+ * A draggable context is used to define an area in the page where drag and drop can be used on
+ * one or more elements (you also need to use WithDrop and WithDrag):
  *
- *   <ol>
- *     <DraggableContext
- *       moveItem={ onItemMoved }
- *       canDrag={ itemCanDrag }
- *       beginDrag={ onBeginDrag }
- *       hover={ onHover }
- *     >
- *       <li>Spring roll</li>
- *       <li>Prawn toast</li>
- *       <li>Hot and sour soup</li>
- *       <li>Spare ribs</li>
- *     </DraggableContext>
- *   </ol>
+ *   <DraggableContext onDrag={ onItemMoved }>
+ *     <ol>
+ *       {
+ *         items.map((item, index) => {
+ *           return (
+ *             <WithDrop index={ index }>
+ *               <li>
+ *                 <WithDrag><span>{ item.content }</span></WithDrag>
+ *               </li>
+ *             </WithDrop>
+ *           );
+ *         })
+ *       }
+ *     </ol>
+ *   </DraggableContext>
  *
  * @class DraggableContext
  * @constructor
@@ -35,49 +39,12 @@ class DraggableContext extends React.Component {
 
   static propTypes = {
     /**
-     * Callback function for when an item has been dragged and dropped
+     * Callback function for when an item has been dragged
      * e.g. to update data in a store
-     *
-     * @property moveItem
-     * @type {Function}
      */
-    moveItem: React.PropTypes.func,
-
-    /**
-     * Callback function that determines whether the item can be
-     * dragged.
-     *
-     * Synonymous with React DnD's Drag Source Specification canDrag
-     * - see https://react-dnd.github.io/react-dnd/docs-drag-source.html
-     *
-     * @property canDrag
-     * @type {Function}
-     */
-    canDrag: React.PropTypes.func,
-
-    /**
-     * Callback function called when dragging starts.
-     *
-     * Synonymous with React DnD's Drag Source Specification beginDrag
-     * - see https://react-dnd.github.io/react-dnd/docs-drag-source.html
-     *
-     * @property beginDrag
-     * @type {Function}
-     */
-    beginDrag: React.PropTypes.func,
-
-    /**
-     * Callback function called when an item is hovered over the
-     * component.
-     *
-     * Synonymous with React DnD's Drop Target Specification
-     * - see https://react-dnd.github.io/react-dnd/docs-drop-target.html
-     *
-     * @property hover
-     * @type {Function}
-     */
-    hover: React.PropTypes.func
+    onDrag: PropTypes.func.isRequired
   }
+
 
   /**
    * Defines a context object for child components of the draggable context component.
@@ -87,32 +54,85 @@ class DraggableContext extends React.Component {
    * @type {Object}
    */
   static childContextTypes = {
-    moveItem: React.PropTypes.func, // See propTypes.moveItem
-    canDrag: React.PropTypes.func, // See propTypes.canDrag
-    beginDrag: React.PropTypes.func, // See propTypes.beginDrag
-    hover: React.PropTypes.func // See propTypes.hover
+    dragAndDropOnDrag: PropTypes.func, // Callback for when order is changed
+    dragAndDropBeginDrag: PropTypes.func, // Callback for when dragging begins
+    dragAndDropEndDrag: PropTypes.func, // Callback for when dragging ends
+    dragAndDropHover: PropTypes.func, // Callback for when a hover is triggered
+    dragAndDropActiveIndex: PropTypes.number // Tracks the currently dragged index
   }
 
   /**
-   * Returns this draggable context object to child components.
+   * Returns this draggable context properties to child components.
    *
    * @method getChildContext
    * @return {void}
    */
   getChildContext() {
     return {
-      moveItem: this.props.moveItem,
-      canDrag: this.props.canDrag,
-      beginDrag: this.props.beginDrag,
-      hover: this.props.hover
+      dragAndDropOnDrag: this.handleDrag,
+      dragAndDropBeginDrag: this.handleBeginDrag,
+      dragAndDropEndDrag: this.handleEndDrag,
+      dragAndDropHover: this.handleHover,
+      dragAndDropActiveIndex: this.state.activeIndex
+    };
+  }
+
+  state = {
+    activeIndex: null // {Number} tracks the currently dragged index
+  }
+
+  /**
+   * A callback for when hover is triggered
+   *
+   * @method Hover
+   * @return {Void}
+   */
+  handleHover = ItemTargetHelper.onHoverUpDown
+
+  /**
+   * A callback for when a drag is triggered.
+   *
+   * @method handleDrag
+   * @param {Number} originalIndex (the active item's original index)
+   * @param {Number} hoverIndex (the active item's new index
+   * @return {Void}
+   */
+  handleDrag = (originalIndex, hoverIndex) => {
+    // tracks the new index
+    this.setState({ activeIndex: hoverIndex });
+
+    // only triggers the onDrag callback if there is an originalIndex
+    if (typeof originalIndex !== 'undefined') {
+      this.props.onDrag(originalIndex, hoverIndex);
+    }
+  }
+
+  /**
+   * A callback for when dragging begins.
+   *
+   * @method handleBeginDrag
+   * @param {Object} props
+   * @return {Void}
+   */
+  handleBeginDrag = (props) => {
+    return {
+      index: props.index
     };
   }
 
   /**
-   * Renders the component
+   * A callback for when a drag ends triggered.
    *
-   * @method render
-   * @return {Object} JSX
+   * @method handleEndDrag
+   * @return {Void}
+   */
+  handleEndDrag = () => {
+    // dragging has ended so remove the active index
+    this.setState({ activeIndex: null });
+  }
+
+  /**
+   * Renders the component
    */
   render() {
     return this.props.children;