+++ CHANGELOG.md
@@ -1,3 +1,53 @@
+# 1.1.0
+
+## DraggableContext, WithDrag & WithDrop
+
+We now provide a series of components to enable drag and drop functionality. For example:
+
+```
+<DraggableContext onDrag={ onItemMoved }>
+  <ol>
+    {
+      items.map((item, index) => {
+        return (
+          <WithDrop key={ index } index={ index }>
+            <li>
+              <WithDrag><span>{ item.content }</span></WithDrag>
+            </li>
+          </WithDrop>
+        );
+      });
+    }
+  </ol>
+</DraggableContext>
+```
+
+The `onDrag` prop to manipulates the order as it changes. It is a function that receives two arguments when called: `dragIndex`, which is the original position of the item, and `hoverIndex`, which is the position of the item if dropped.
+
+An example function signature: `onItemMoved = (dragIndex, hoverIndex) => { }`
+
+## Draggable Table Rows
+
+The `TableRow` component now supports drag and drop. To enable it you need to add a `DraggableContext` component, apply an index to each `TableRow`, and define the `onDrag` prop to manipulate the order as it changes:
+
+```
+<Table tbody={ false }>
+  <DraggableContext onDrag={ onRowMoved }>
+    <tbody>
+      {
+        rows.map((row, index) => {
+          return (
+            <TableRow key={ index } index={ index }>
+              { row.content }
+            </TableRow>
+          );
+        });
+      }
+    </tbody>
+  </DraggableContext>
+</Table>
+```
+
 # 1.0.0
 
 ## :warning: Major Change - React 15 Upgrade